requirements.txt
customtkinter==5.2.2
Pillow==12.0.0
mysql-connector-python==9.5.0
matplotlib==3.10.7
fpdf2==2.8.5
